backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-bigquery==3.27.0
pydantic==2.10.4
